% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PASCCA.R
\name{PAlist}
\alias{PAlist}
\title{Produce poly(A) list}
\usage{
PAlist(data)
}
\arguments{
\item{data}{Either a numeric matrix or a data frame.
The expressions of RNA-seq data at poly(A) levels,
exon levels or position levels.}
}
\value{
PAlist returns an object of poly(A) list.
}
\description{
This function can produce a poly(A) list.
}
