% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PASCCA.R
\name{PAlog}
\alias{PAlog}
\title{Data transformation}
\usage{
PAlog(data)
}
\arguments{
\item{data}{Either a numeric matrix or a data frame.
The expressions of RNA-seq data at poly(A) levels,
exon levels or position levels.}
}
\value{
An object of expression matrix or data frame
    the same as the input data.
}
\description{
To transform the count data to the log2 scale.
}
