% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PASCCA.R
\name{PAprocess}
\alias{PAprocess}
\title{APA-related gene expression data preprocessing}
\usage{
PAprocess(data, log = TRUE)
}
\arguments{
\item{data}{Either a numeric matrix or a data frame.The expressions
of RNA-seq data  at exon levels or position levels. The first column
is poly(A) or exon names, the second column is gene names, and the
remaining column is sample names under different biological conditions
such as different tissues, cell types and developmental stages.
If the set of samples have repeated measurements, the order of the
samples in data must be arranged in from big to small according to
the number of replicates.}

\item{log}{Choose Whether or not to perform log2 transformation for data.}
}
\value{
An object of expression matrix or data frame the same as the input
    data.
}
\description{
It鈥檚 to do preliminary processing of APA-related gene
    expression data.User can choose whether to perform log-transformation
    by parameter log (default: TRUE).
}
\details{
This function provides two steps taken to pre-process data. 1) Data cleaning:
To filter out genes with one poly(A) site from gene expression data. 2) Data
transformation: To transform the count data to the log2 scale by setting the
parameter log=TRUE (default: TRUE).
}
\note{
If the set of samples have repeated measurements, the order of the
    samples in data must be arranged in from big to small according to
    the number of replicates.
}
\examples{
 ##Loading example data
 data(polyA_example_data2)
 dim(data2)
 str(data2)
 data2[1:5,1:5]

 ##Data preprocessing
 pre_data <- PAprocess(data2,log="T")
 dim(pre_data)
 pre_data[1:5,1:5]
}
\references{
Thomas, P. E., X. Wu, M. Liu, B. Gaffney, G. Ji et al.,
    2012 Genome-Wide Control of Polyadenylation Site Choice by CPSF30
    in Arabidopsis. Plant Cell 24: 4376-4388.
}
\author{
Yuqi Long, Wenbin Ye
}
