% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PASCCA.R
\name{SCCAdist}
\alias{SCCAdist}
\title{Compute shrinkage correlation coefficient distance}
\usage{
SCCAdist(data)
}
\arguments{
\item{data}{Either a numeric matrix or a data frame.
The expressions of RNA-seq data at poly(A) levels,
exon levels or position levels.}
}
\value{
A shrinkage correlation coefficient distance
matrix.
}
\description{
To compute shrinkage correlation coefficient distance
matrix by parallel.
}
